% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv-grpsel.R
\name{coef.cv.grpsel}
\alias{coef.cv.grpsel}
\title{Coefficient function for cv.grpsel object}
\usage{
\method{coef}{cv.grpsel}(object, lambda = "lambda.min", gamma = "gamma.min", ...)
}
\arguments{
\item{object}{an object of class \code{cv.grpsel}}

\item{lambda}{the value of \code{lambda} indexing the desired fit}

\item{gamma}{the value of \code{gamma} indexing the desired fit}

\item{...}{any other arguments}
}
\value{
A matrix of coefficients.
}
\description{
Extracts coefficients for specified values of the tuning parameters.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
