% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpsel-cv.R
\name{plot.cv.grpsel}
\alias{plot.cv.grpsel}
\title{Plot function for cv.grpsel object}
\usage{
\method{plot}{cv.grpsel}(x, gamma = "gamma.min", ...)
}
\arguments{
\item{x}{an object of class \code{cv.grpsel}}

\item{gamma}{the value of gamma indexing the desired fit}

\item{...}{any other arguments}
}
\value{
A plot of the cross-validation results.
}
\description{
Plot the cross-validation results from group subset selection for a specified value
of \code{gamma}.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
