% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpsel.R
\name{plot.grpsel}
\alias{plot.grpsel}
\title{Plot function for grpsel object}
\usage{
\method{plot}{grpsel}(x, gamma = 0, ...)
}
\arguments{
\item{x}{an object of class \code{grpsel}}

\item{gamma}{the value of \code{gamma} indexing the desired fit}

\item{...}{any other arguments}
}
\value{
A plot of the coefficient profiles.
}
\description{
Plot the coefficient profiles from group subset selection for a specified value of
\code{gamma}.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
