% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpsel.R
\name{predict.grpsel}
\alias{predict.grpsel}
\title{Predict function for grpsel object}
\usage{
\method{predict}{grpsel}(object, x.new, lambda = NULL, gamma = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{grpsel}}

\item{x.new}{a matrix or array of new values for the predictors}

\item{lambda}{the value of lambda indexing the desired fit}

\item{gamma}{the value of gamma indexing the desired fit}

\item{...}{any other arguments}
}
\value{
A matrix or array of predictions.
}
\description{
Generate predictions for new data using specified values of the tuning parameters.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
