% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpsel.R
\name{coef.grpsel}
\alias{coef.grpsel}
\title{Coefficient function for grpsel object}
\usage{
\method{coef}{grpsel}(object, lambda = NULL, gamma = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{grpsel}}

\item{lambda}{the value of \code{lambda} indexing the desired fit}

\item{gamma}{the value of \code{gamma} indexing the desired fit}

\item{...}{any other arguments}
}
\value{
A matrix of coefficients.
}
\description{
Extracts coefficients for specified values of the tuning parameters.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
