% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grpsel-cv.R
\name{predict.cv.grpsel}
\alias{predict.cv.grpsel}
\title{Predict function for cv.grpsel object}
\usage{
\method{predict}{cv.grpsel}(object, x.new, lambda = "lambda.min", gamma = "gamma.min", ...)
}
\arguments{
\item{object}{an object of class \code{cv.grpsel}}

\item{x.new}{a matrix or array of new values for the predictors}

\item{lambda}{the value of lambda indexing the desired fit}

\item{gamma}{the value of gamma indexing the desired fit}

\item{...}{any other arguments}
}
\value{
A matrix or array of predictions.
}
\description{
Generate predictions for new data using specified values of the tuning parameters.
}
\author{
Ryan Thompson <ryan.thompson@monash.edu>
}
